rewrite all functions to follow the same style, move modals components to seperate folder
