+++ frontend/src/components/modals/BuildingMapModal.tsx
