+++ frontend/src/pages/Login.tsx
@@ -1,4 +1,3 @@
-
 import { GoogleLogin } from "@react-oauth/google";
 import { useNavigate, useLocation } from "react-router-dom";
 import { api } from "@services/axios";
